requirements.txt
openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
mautrix>=0.19,<0.20
python-magic
libmagic
lxml
beautifulsoup4